# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenRelay API",
    "description": "Form orchestration backend.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "list_users",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/select-list": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "select_list",
        "responses": {
          "200": {
            "description": "Dropdown-friendly user list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSelectOption"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission, self-delete, or last-superadmin"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "operationId": "update_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission or last-superadmin demotion"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{id}/password": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "change_password",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSetPassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password reset; target's sessions revoked"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission or target outranks actor"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AdminSetPassword": {
        "type": "object",
        "description": "Admin-initiated password reset for another user. No current-password proof\n(the actor is acting by permission, not as the target), but the service\nforbids resetting a user who outranks the actor and revokes the target's\nsessions. See `service::admin_set_password`.",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "NewUser": {
        "type": "object",
        "description": "Input shape for creating a user. `role_ids` is optional and defaults to\nan empty set — the route handler reads it and calls into the RBAC service\nin the same transaction after the user row is inserted.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "RoleSummary": {
        "type": "object",
        "description": "Lightweight role reference — enough to render a badge, populate a\ndropdown, or describe the current user's roles in `/auth/me`. Cheap to\nembed in `UserDto` since no permission join is required.",
        "required": [
          "id",
          "name",
          "is_system"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "is_system": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateUser": {
        "type": "object",
        "description": "Partial update. `None` means \"leave the field alone\". For `display_name`,\nan explicit empty string is treated as \"clear\".\n\n`role_ids` semantics are asymmetric vs `display_name`: `None` leaves\nassignments alone, `Some(vec![])` clears all roles, `Some(non-empty)`\nreplaces the set. The role-assignment guard (last-superadmin) lives in\n`crate::rbac::service::assign_roles_to_user`.",
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "role_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "description": "Outbound representation of a user — what callers see in API responses.\n\n`roles` is populated by `service::populate_roles` (or by the RBAC service\ndirectly for single-user lookups). `From<user::Model>` produces an empty\nroles vec; the responsibility to enrich falls to the caller so this type\nstays cheap to construct in places (login, setup) where the frontend will\nrefetch its session shape via `/auth/me` anyway.",
        "required": [
          "id",
          "email"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            }
          }
        }
      },
      "UserList": {
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "UserSelectOption": {
        "type": "object",
        "description": "Lightweight option used to populate dropdowns elsewhere in the admin.",
        "required": [
          "id",
          "label"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "label": {
            "type": "string"
          }
        }
      }
    }
  },
...
}